+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 1.0.0.beta.2
 
 * Addition of URL validator.
+* Addition of email validator.